requirements.txt
google-genai>=0.1.0
PyYAML>=6.0
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
gitpython>=3.1.0